pyproject.toml
[project]
name = "quora-analysis"
version = "0.1.0"
description = "Quora Answer Scraper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "scrapy>=2.13.2",
    "psycopg2-binary>=2.9.9",
    "selenium>=4.15.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "html2text>=2020.1.16",
    "fake-useragent>=1.4.0",
    "webdriver-manager>=4.0.0",
    "html-to-markdown>=1.4.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
